chore(reports): slice 9 — remove old reports UI + update architecture doc
Final cleanup. The new chat-as-timeline builder + library replace the
single-page ReportsView entirely.

Removed
- src/features/reports/components/ReportsView.tsx (the old single-page
  layout with sidebar + chat + results)
- src/features/reports/components/ChatPanel.tsx (used only by ReportsView)
- src/features/reports/components/SavedReportsSidebar.tsx (replaced by
  the new /reports library page)
- src/features/reports/hooks/useSavedReports.ts (replaced by lib/queries.ts;
  useRunSavedReport moved over before deletion)
- src/features/reports/components/__tests__/ChatPanel.test.tsx
- The window.prompt save flow that lived inside ReportsView (replaced by
  the new SavePopover in slice 6)

Moved
- useRunSavedReport now lives in src/features/reports/lib/queries.ts so
  the new builder doesn't reach back into the deleted hooks file.

Documented
- docs/architecture.md: added the reports feature row to the Feature
  Directory Map, added a 'reports' Cross-Feature Dependencies entry
  pointing at the readonly Postgres pool + agent metadata module, and
  added a full 'The reports Feature' section enumerating URL state
  semantics and the components/hooks/lib tree.

Verification: 191 tests across 30 files green; tsc clean.
